new page and edit page split
separated edit page and new page into two different files

separated the form into a separate file
