+++ src/index.tsx
@@ -7,6 +7,7 @@ import { BrowserRouter, Routes, Route, Navigate } from "react-router-dom";
 import ReadingPages from "./ReadingPages";
 import CookingPages from "./CookingPages";
 import NewRecipe from "./NewRecipePage";
+import EditRecipe from "./EditRecipePage";
 
 const root = ReactDOM.createRoot(
   document.getElementById("root") as HTMLElement
@@ -20,7 +21,7 @@ root.render(
         <Route path="reading/:recipeId" element={<ReadingPages />}></Route>
         <Route path="cooking/:recipeId" element={<CookingPages />}></Route>
         <Route path="newrecipe" element={<NewRecipe />}></Route>
-        <Route path="editrecipe/:recipeId" element={<NewRecipe />}></Route>
+        <Route path="editrecipe/:recipeId" element={<EditRecipe />}></Route>
         <Route
           path="*"
           element={
